Move reason message handle from [initState] to [askingReplyRequestReason]
