+++ src/webhook/handlers/initState.js
@@ -1,16 +1,12 @@
 import stringSimilarity from 'string-similarity';
-import { t, msgid, ngettext } from 'ttag';
+import { t } from 'ttag';
 import gql from 'src/lib/gql';
-import { REASON_PREFIX, getArticleURL } from 'src/lib/sharedUtils';
 import {
-  ManipulationError,
   createPostbackAction,
   ellipsis,
   createAskArticleSubmissionConsentReply,
   createSuggestOtherFactCheckerReply,
   POSTBACK_NO_ARTICLE_FOUND,
-  createReasonButtonFooter,
-  createArticleShareBubble,
   createHighlightContents,
 } from './utils';
 import ga from 'src/lib/ga';
@@ -20,98 +16,6 @@ const SIMILARITY_THRESHOLD = 0.95;
 export default async function initState(params) {
   let { data, event, userId, replies, isSkipUser } = params;
   let state = '__INIT__';
-
-  if (event.input.startsWith(REASON_PREFIX)) {
-    // Check required data to update reply request
-    if (!data.selectedArticleId) {
-      throw new ManipulationError(
-        t`Please press the latest button to submit message to database.`
-      );
-    }
-
-    // Update the reply request
-    const { data: mutationData, errors } = await gql`
-      mutation UpdateReplyRequest($id: String!, $reason: String) {
-        CreateOrUpdateReplyRequest(articleId: $id, reason: $reason) {
-          text
-          replyRequestCount
-        }
-      }
-    `(
-      {
-        id: data.selectedArticleId,
-        reason: event.input.slice(REASON_PREFIX.length),
-      },
-      { userId }
-    );
-
-    if (errors) {
-      throw new ManipulationError(
-        t`Something went wrong when recording your reason, please try again later.`
-      );
-    }
-
-    const visitor = ga(
-      userId,
-      state,
-      mutationData.CreateOrUpdateReplyRequest.text
-    );
-    visitor.event({
-      ec: 'Article',
-      ea: 'ProvidingReason',
-      el: data.selectedArticleId,
-    });
-
-    const articleUrl = getArticleURL(data.selectedArticleId);
-    const otherReplyRequestCount =
-      mutationData.CreateOrUpdateReplyRequest.replyRequestCount - 1;
-    const replyRequestUpdatedMsg = t`Thanks for the info you provided.`;
-
-    replies = [
-      {
-        type: 'flex',
-        altText: replyRequestUpdatedMsg,
-        contents: {
-          type: 'carousel',
-          contents: [
-            createArticleShareBubble(articleUrl),
-            {
-              type: 'bubble',
-              body: {
-                type: 'box',
-                layout: 'vertical',
-                contents: [
-                  {
-                    type: 'text',
-                    wrap: true,
-                    text: replyRequestUpdatedMsg,
-                  },
-                  otherReplyRequestCount > 0 && {
-                    type: 'text',
-                    wrap: true,
-                    text: ngettext(
-                      msgid`There is ${otherReplyRequestCount} user also waiting for clarification.`,
-                      `There are ${otherReplyRequestCount} users also waiting for clarification.`,
-                      otherReplyRequestCount
-                    ),
-                  },
-                ].filter(m => m),
-              },
-              footer: createReasonButtonFooter(
-                articleUrl,
-                userId,
-                data.sessionId,
-                true
-              ),
-            },
-          ],
-        },
-      },
-    ];
-    visitor.send();
-
-    return { data, event, userId, replies, isSkipUser };
-  }
 
   // Track text message type send by user
   const visitor = ga(userId, state, event.input);
